updated OperatorOverloadingTrial1 file
Operator overloading successfully done to store the entered key and value in the tempkey and tempval variables of that object. used a const reference to take in the rvalue into a variable(which automatically creates a temporary variable in memory, hence making that string literal act like an lvalue). Then I simply stored this->tempval in a variable called valref and returned valref( so that I return an lvalue) and the rest of the job was done by the assignment operator. i.e. the LHS of agemap[sheldon] = 25 is nothing but valref which is an lvalue. So 25 then gets assigned to valref