added simplified_unordered_map
simplified_unordered_map uses the same algorithm as the real thing, but I relieved it of a lot of technicals to focus on the algorithm. the algorithm works and it successfully handles collisions and searches through the linked lists. now I need to apply the same algorithm to the actual big program, with some tweaks so that it fits into the operator overloading stuff and automatically stores the keys and values that the user declairs