unordered_map_v1_complete
unordered_map is working. It can store and return values in O(1) complexity and if a new assignment is made to the same key, the first one will automatically be deleted. Using malloc for creating a new node was apparently resulting in a seg fault; Used the new keyword instead